+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Fixes
 
+* Fix a Webpack cache issue leading to modules symlinked in `node_modules` not being rebuilt. 
 * Fixes login maximum attempts error message that wasn't showing the plural when lockoutMinutes is more than 1.
 * Fixes the text color of the current array item's slat label in the array editor modal 
 * Fixes the maximum width of an array item's slat label so as to not obscure the Remove button in narrow viewports